Rename WorkflowHistoryVersionDot to WorkflowVersionStatusIndicator
